client/requirements.txt
furhat-realtime-api
google-genai